fullstack-todo/chatbot_backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "AI Chatbot Integration for Todo Application"
authors = [{name = "Developer", email = "dev@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "openai>=1.12.0",
    "sqlmodel>=0.0.16",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0",
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "psycopg[binary]>=3.1.18",
    "alembic>=1.13.1",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.1.0",
    "redis>=5.0.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["chatbot_backend*"]

[tool.uvicorn]
host = "0.0.0.0"
port = 8001
reload = true
